chore: move TestxServer class in separate file
